refactor: extract sandbox compiler and DRY CLI evaluation
Move sandbox-specific compilation logic (_compile_sandbox, _extract_paths,
_extract_command, _extract_urls, _extract_hostname, _domain_matches) from
compiler.py into a dedicated sandbox_compiler.py module. This reduces
compiler.py from 611 to 432 lines and gives sandbox compilation a focused
home.

Also consolidate the repeated precondition+sandbox evaluation pattern in
CLI commands (check, replay, test) into a single _evaluate_all() helper,
eliminating 3 duplicated blocks.

No behavioral changes — pure structural refactoring. 2043 tests pass.
